+++ react/src/components/HeaderMenu.tsx
@@ -54,7 +54,6 @@ const MenuLinks = ({ isOpen, onLinkClick, ...props }) => {
         <MenuItem to="/" text="Home" />
         <MenuItem to="/locations" text="Locations" />
         <MenuItem to="/boxes" text="Boxes" />
-        <MenuItem to="/table" text="Table" />
       </Stack>
     </Box>
   );
