+++ react/src/views/boxes/GlobalFilter.tsx
@@ -3,7 +3,8 @@ import { useAsyncDebounce } from "react-table";
 import { SearchIcon } from "@chakra-ui/icons";
 import { Input, InputGroup, InputRightElement } from "@chakra-ui/react";
 
-export function GlobalFilter({ preGlobalFilteredRows, globalFilter, setGlobalFilter }) {
+export function GlobalFilter({ globalFilter, setGlobalFilter }) {
+  console.log("globalFilter", globalFilter);
   const [value, setValue] = useState(globalFilter);
   const onChange = useAsyncDebounce((value) => {
     setGlobalFilter(value || undefined);
